chore(__tests__): make a function to mock fastify server build in tests
