+++ packages/api-gateway/mocks/geolocation.mock.js
@@ -1,4 +1,8 @@
 const { faker } = require('@faker-js/faker');
+const Fastify = require('fastify');
+const Autoload = require('fastify-autoload');
+const pluginLoader = require('fastify-plugin');
+const getApiGatewayDirectory = require('../src/utils/getApiGatewayDirectory');
 
 // eslint-disable-next-line no-unused-vars
 function mockPlaces() {
@@ -48,4 +52,19 @@ function mockDeletePlace(id) {
   return place;
 }
 
-module.exports = { mockPlaces, mockAddress, mockDeletePlace };
+function mockBuildApp(GeolocationServices) {
+  async function services(fastify) {
+    await fastify.decorate('geolocationServices', GeolocationServices);
+  }
+
+  // load fastify and update function
+  const fastify = Fastify({ logger: true });
+  fastify.register(Autoload, { dir: getApiGatewayDirectory() });
+  fastify.register(pluginLoader(services));
+
+  return fastify;
+}
+
+module.exports = {
+  mockPlaces, mockAddress, mockDeletePlace, mockBuildApp,
+};
